analysis/src/requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
idna==2.10
# git+https://github.com/addisonlynch/iexfinance@b23c0646f614c84dc32750aee3bbb3cf10afbe35#egg=iexfinance
iexfinance==0.5.0
tables==3.6.1
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-cache==0.5.2
six==1.15.0
urllib3==1.25.10
